requirements.txt
streamlit>=1.36
opencv-python>=4.10
numpy>=2.0
pandas>=2.2
scikit-learn>=1.5
Pillow>=10.3
matplotlib>=3.9
joblib>=1.4
psutil>=5.9

# PyTorch CPU build for Python 3.14
torch>=2.11
torchvision>=0.20

yt-dlp>=2024.4.9
pytest>=7.4
